Issue 5472/1: yaffut.hh: Fix compilation warning
(cherry picked from commit ddef2aa7b6d285f37983f3e7ca17dde5436221bc)

+++ flower/include/yaffut.hh
@@ -209,7 +209,7 @@ public:
       }
 
     Factory::Instance ().Report ();
-    return Factory::Instance ().Fail ();
+    return static_cast<int> (Factory::Instance ().Fail ());
   }
 };
